Fix logging in the tests
Now we get a nice and proper traceback from server errors and the server
log for failed tests (and for passing tests as well, if one wants to).

I don't think® this makes a difference, the configuration in the alembic.ini
was pretty similar to the default configuration. The server run via
main() overwrites the logging configuration with the one from --logging-config